requirements.txt
absl-py==0.11.0
astunparse==1.6.3
attrs==20.3.0
backcall==0.2.0
cachetools==4.2.0
chardet==4.0.0
coverage==5.3.1
cycler==0.10.0
decorator==4.4.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.4.0
iniconfig==1.1.1
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.0
jsonpatch==1.28
jsonpointer==2.0
keras-preprocessing==1.1.2
kiwisolver==1.3.1
lmdb==1.0.0
markdown==3.3.3
matplotlib==3.3.3
networkx==2.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.10
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygments==2.7.4
pyparsing==2.4.7
pytest==6.2.1
pytest-cov==2.10.1
python-dateutil==2.8.1
pytz==2020.5
pywavelets==1.1.1
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.18.1
scikit-learn==0.24.0
scipy==1.6.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.1.11
toml==0.10.2
torch-mimicry==0.1.16
torchfile==0.1.0
tornado==6.1
tqdm==4.56.1
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
websocket-client==0.57.0
werkzeug==1.0.1
wrapt==1.12.1
yapf==0.30.0
zipp==3.4.0
